// src/types/interface/server-list-item.interface.ts
export default interface ServerListItem{
    serverId: number;
    name: string;
    content: string;
    location: string;
    performance: string;
    disk: string;
    backup: boolean;
    billingAmount: string;
    requestDetails: string;
    modeCount: number;
    gameTitle: string;
    gameImage: string;
    serverUserId: string;
    status: string;
    serverAddress: string;
    userId: string;
}